+++ Components/IRacing/Events/IRacingCompletedLap.cs
@@ -10,7 +10,7 @@ namespace Slipstream.Components.IRacing.Events
         public ulong Uptime { get; set; }
         public double SessionTime { get; set; }
         public long CarIdx { get; set; }
-        public double Time { get; set; }
+        public double LapTime { get; set; }
         public int LapsCompleted { get; set; }
         public float? FuelDelta { get; set; }
         public bool LocalUser { get; set; }
@@ -23,7 +23,7 @@ namespace Slipstream.Components.IRacing.Events
                    ExcludeFromTxrx == lap.ExcludeFromTxrx &&
                    SessionTime == lap.SessionTime &&
                    CarIdx == lap.CarIdx &&
-                   Time == lap.Time &&
+                   LapTime == lap.LapTime &&
                    LapsCompleted == lap.LapsCompleted &&
                    FuelDelta == lap.FuelDelta &&
                    LocalUser == lap.LocalUser &&
@@ -37,7 +37,7 @@ namespace Slipstream.Components.IRacing.Events
             hashCode = hashCode * -1521134295 + ExcludeFromTxrx.GetHashCode();
             hashCode = hashCode * -1521134295 + SessionTime.GetHashCode();
             hashCode = hashCode * -1521134295 + CarIdx.GetHashCode();
-            hashCode = hashCode * -1521134295 + Time.GetHashCode();
+            hashCode = hashCode * -1521134295 + LapTime.GetHashCode();
             hashCode = hashCode * -1521134295 + LapsCompleted.GetHashCode();
             hashCode = hashCode * -1521134295 + FuelDelta.GetHashCode();
             hashCode = hashCode * -1521134295 + LocalUser.GetHashCode();